refactor(resources): move threads resource into resources package
